Manual 2/requirement_1/MNIST Digit Recognition Project/requirements.txt
contourpy
cycler
fonttools
joblib
kiwisolver
matplotlib
numpy
opencv-python
packaging
pandas
pillow
pyparsing
python-dateutil
pytz
scikit-learn
scipy
six
threadpoolctl
tzdata
gdown
seaborn
scikit-image
requests
zipfile

